requirements.txt
# Core Dependencies
python-dotenv==1.0.0           # Environment variable management
aiohttp==3.9.1                 # Async HTTP client/server
websockets==12.0               # WebSocket client/server
pymongo==4.6.1                 # MongoDB async driver
motor==3.3.2                   # Async MongoDB driver

# Twitter Integration
twitter-api-client==0.10.3     # Twitter API integration

# AI/ML
openai==1.3.7                  # OpenAI/ChatGPT API client
tiktoken==0.5.2               # OpenAI token counting

# CLI and Interface
rich==13.7.0                   # Terminal formatting and UI
prompt-toolkit==3.0.43         # Interactive command line interface
typer==0.9.0                   # CLI builder

# Async Support
asyncio==3.4.3                 # Async I/O support
aiofiles==23.2.1              # Async file operations

# Utilities
pydantic==2.5.2               # Data validation
loguru==0.7.2                 # Advanced logging
schedule==1.2.1               # Task scheduling
tenacity==8.2.3               # Retry handling

# Development Tools
black==23.11.0                # Code formatting
isort==5.12.0                 # Import sorting
mypy==1.7.1                   # Static type checking
pytest==7.4.3                 # Testing framework
pytest-asyncio==0.23.2        # Async testing support
pytest-cov==4.1.0             # Test coverage

# Documentation
mkdocs==1.5.3                 # Documentation generator
mkdocs-material==9.5.2        # Documentation theme
